Restructure java classes into different packages + set title
